% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\name{add_constraint}
\alias{add_constraint}
\title{Add a constraint to the model}
\usage{
add_constraint(model, expr, ...)
}
\arguments{
\item{model}{the model}

\item{expr}{an expression}

\item{...}{any quantifiers}
}
\description{
Add a constraint to the model
}
