% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\name{add_variable}
\alias{add_variable}
\title{Add a variable to the model}
\usage{
add_variable(model, expr, type = "continuous", lb = -Inf, ub = Inf,
  ...)
}
\arguments{
\item{model}{the model}

\item{expr}{an expression}

\item{type}{a character, either continuous, integer or binary}

\item{lb}{the lower bound}

\item{ub}{the upper bound}

\item{...}{any quantifiers}
}
\description{
Add a variable to the model
}
