% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{==,MOI_abstract_function,numeric-method}
\alias{==,MOI_abstract_function,numeric-method}
\title{eq relation}
\usage{
\S4method{==}{MOI_abstract_function,numeric}(e1, e2)
}
\arguments{
\item{e1}{a MOI_abstract_function}

\item{e2}{a numeric}
}
\description{
eq relation
}
\keyword{internal}
