% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{==,numeric,MOI_abstract_function-method}
\alias{==,numeric,MOI_abstract_function-method}
\title{eq relation}
\usage{
\S4method{==}{numeric,MOI_abstract_function}(e1, e2)
}
\arguments{
\item{e1}{a numeric}

\item{e2}{a MOI_abstract_function}
}
\description{
eq relation
}
\keyword{internal}
