% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{<=,numeric,MOI_abstract_function-method}
\alias{<=,numeric,MOI_abstract_function-method}
\title{leq relation}
\usage{
\S4method{<=}{numeric,MOI_abstract_function}(e1, e2)
}
\arguments{
\item{e1}{a MOI_abstract_function}

\item{e2}{a numeric}
}
\description{
leq relation
}
