% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\name{Model}
\alias{Model}
\title{A new mixed-integer programming Model}
\usage{
Model(solver = cache_solver())
}
\arguments{
\item{solver}{a solver object}
}
\description{
A new mixed-integer programming Model
}
