% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{-,MOI_scalar_quadratic_function,MOI_scalar_affine_term-method}
\alias{-,MOI_scalar_quadratic_function,MOI_scalar_affine_term-method}
\title{Substract a scalar affine term and a scalar quadratic function}
\usage{
\S4method{-}{MOI_scalar_quadratic_function,MOI_scalar_affine_term}(e1, e2)
}
\arguments{
\item{e1}{A scalar quadratic function}

\item{e2}{A scalar affine term}
}
\description{
Substract a scalar affine term and a scalar quadratic function
}
