% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{-,MOI_scalar_quadratic_term,MOI_scalar_affine_function-method}
\alias{-,MOI_scalar_quadratic_term,MOI_scalar_affine_function-method}
\title{Substract a scalar affine function and a scalar quadratic term}
\usage{
\S4method{-}{MOI_scalar_quadratic_term,MOI_scalar_affine_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar quadratic term}

\item{e2}{A scalar affine function}
}
\description{
Substract a scalar affine function and a scalar quadratic term
}
\keyword{internal}
