% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{-,numeric,MOI_scalar_affine_term-method}
\alias{-,numeric,MOI_scalar_affine_term-method}
\title{Substract a scalar affine term and a numeric}
\usage{
\S4method{-}{numeric,MOI_scalar_affine_term}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric}

\item{e2}{A scalar affine term}
}
\description{
Substract a scalar affine term and a numeric
}
\keyword{internal}
