% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{+,MOI_scalar_affine_term,missing-method}
\alias{+,MOI_scalar_affine_term,missing-method}
\title{Unary plus of a scalar affine term}
\usage{
\S4method{+}{MOI_scalar_affine_term,missing}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine term}

\item{e2}{missing}
}
\description{
Unary plus of a scalar affine term
}
\keyword{internal}
