% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-expression-methods.R
\name{+,MOI_scalar_affine_term,MOI_scalar_affine_function-method}
\alias{+,MOI_scalar_affine_term,MOI_scalar_affine_function-method}
\title{Add a scalar affine function and a scalar affine term}
\usage{
\S4method{+}{MOI_scalar_affine_term,MOI_scalar_affine_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine term}

\item{e2}{A scalar affine function}
}
\description{
Add a scalar affine function and a scalar affine term
}
\keyword{internal}
