% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{+,MOI_scalar_affine_term,MOI_scalar_affine_term-method}
\alias{+,MOI_scalar_affine_term,MOI_scalar_affine_term-method}
\title{Add a scalar affine term and a scalar affine term}
\usage{
\S4method{+}{MOI_scalar_affine_term,MOI_scalar_affine_term}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine term}

\item{e2}{A scalar affine term}
}
\description{
Add a scalar affine term and a scalar affine term
}
