% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{+,MOI_scalar_affine_term,numeric-method}
\alias{+,MOI_scalar_affine_term,numeric-method}
\title{Add a scalar affine term and a numeric}
\usage{
\S4method{+}{MOI_scalar_affine_term,numeric}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine term}

\item{e2}{A scalar numeric}
}
\description{
Add a scalar affine term and a numeric
}
\keyword{internal}
