% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{+,MOI_scalar_quadratic_function,MOI_scalar_quadratic_function-method}
\alias{+,MOI_scalar_quadratic_function,MOI_scalar_quadratic_function-method}
\title{Add a scalar quadratic function and a scalar quadratic function}
\usage{
\S4method{+}{MOI_scalar_quadratic_function,MOI_scalar_quadratic_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar quadratic function}

\item{e2}{A scalar quadratic function}
}
\description{
Add a scalar quadratic function and a scalar quadratic function
}
