% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-expression-methods.R
\name{+,numeric,MOI_scalar_affine_function-method}
\alias{+,numeric,MOI_scalar_affine_function-method}
\title{Add a scalar affine function and a numeric}
\usage{
\S4method{+}{numeric,MOI_scalar_affine_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric}

\item{e2}{A scalar affine function}
}
\description{
Add a scalar affine function and a numeric
}
