% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{+,numeric,MOI_scalar_quadratic_term-method}
\alias{+,numeric,MOI_scalar_quadratic_term-method}
\title{Add a scalar quadratic term and a numeric}
\usage{
\S4method{+}{numeric,MOI_scalar_quadratic_term}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric}

\item{e2}{A scalar quadratic term}
}
\description{
Add a scalar quadratic term and a numeric
}
