% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-expression-methods.R
\name{^,MOI_scalar_affine_function,numeric-method}
\alias{^,MOI_scalar_affine_function,numeric-method}
\title{Exponentiate a scalar affine function by a numeric}
\usage{
\S4method{^}{MOI_scalar_affine_function,numeric}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine function}

\item{e2}{A numeric of value 2, anything else is not supported}
}
\description{
Exponentiate a scalar affine function by a numeric
}
\keyword{internal}
