% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{^,MOI_scalar_affine_term,numeric-method}
\alias{^,MOI_scalar_affine_term,numeric-method}
\title{Exponentiate a scalar affine term by 2}
\usage{
\S4method{^}{MOI_scalar_affine_term,numeric}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine term}

\item{e2}{A scalar numeric of value 2}
}
\description{
Exponentiate a scalar affine term by 2
}
