% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\docType{data}
\name{RlpMipModel}
\alias{RlpMipModel}
\title{#' Add a variable to the model
#'
#' @param model the model
#' @param expr an expression
#' @param type a character, either continuous, integer or binary
#' @param lb the lower bound
#' @param ub the upper bound
#' @param ... any quantifiers
#'
#' @export
add_variable <- function(model, expr, type = "continuous", lb = -Inf, ub = Inf, ...) {
  rlang::eval_tidy(
    rlang::quo(model$add_variable(!!rlang::enquo(expr), !!type, !!lb, !!ub, ...))
  )
  model
}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RlpMipModel
}
\description{
#' Set the objective of the model
#'
#' @param model the model
#' @param expr an expression
#' @param sense either max or min
#'
#' @export
set_objective <- function(model, expr, sense = "min") {
  rlang::eval_tidy(
    rlang::quo(model$set_objective(!!rlang::enquo(expr), !!sense))
  )
  model
}
}
\details{
#' Add a constraint to the model
#'
#' @param model the model
#' @param expr an expression
#' @param ... any quantifiers
#'
#' @export
add_constraint <- function(model, expr, ...) {
  rlang::eval_tidy(
    rlang::quo(model$add_constraint(!!rlang::enquo(expr), ...))
  )
  model
}

#' Set bounds of individual variables
#'
#' @param model the model
#' @param expr an expression
#' @param lb the lower bounds
#' @param ub the upper bounds
#' @param ... any quantifiers
#'
#' @export
set_bounds <- function(model, expr, lb = NULL, ub = NULL, ...) {
  rlang::eval_tidy(
    rlang::quo(model$set_bounds(!!rlang::enquo(expr), !!lb, !!ub, ...))
  )
  model
}
}
\keyword{datasets}
