% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{RMPK_constraint-class}
\alias{RMPK_constraint-class}
\title{A constraint}
\description{
A constraint
}
\section{Slots}{

\describe{
\item{\code{fun}}{a MOI_abstract_function}

\item{\code{set}}{a MOI_abstract set}
}}

