% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi-solver.R
\name{ROI_solver}
\alias{ROI_solver}
\title{A ROI solver}
\usage{
ROI_solver(solver, control = list())
}
\arguments{
\item{solver}{the solver name}

\item{control}{a list of control options}
}
\description{
A ROI solver
}
