% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\name{set_bounds}
\alias{set_bounds}
\title{Set bounds of individual variables}
\usage{
set_bounds(model, expr, lb = NULL, ub = NULL, ...)
}
\arguments{
\item{model}{the model}

\item{expr}{an expression}

\item{lb}{the lower bounds}

\item{ub}{the upper bounds}

\item{...}{any quantifiers}
}
\description{
Set bounds of individual variables
}
