% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-model.R
\name{set_objective}
\alias{set_objective}
\title{Set the objective of the model}
\usage{
set_objective(model, expr, sense = "min")
}
\arguments{
\item{model}{the model}

\item{expr}{an expression}

\item{sense}{either max or min}
}
\description{
Set the objective of the model
}
