% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-methods.R
\name{[,RMPK_variable_list,ANY,ANY,missing-method}
\alias{[,RMPK_variable_list,ANY,ANY,missing-method}
\title{Access a variable list}
\usage{
\S4method{[}{RMPK_variable_list,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a variable list}

\item{i}{First index}

\item{j}{Second index}

\item{...}{More indexes}

\item{drop}{missing}
}
\description{
Access a variable list
}
\keyword{internal}
