% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-expr.R
\name{sum_expr}
\alias{sum_expr}
\title{Helper function to sum over a set of indexes}
\usage{
sum_expr(expr, ...)
}
\arguments{
\item{expr}{A numeric, linear or quadratic expression}

\item{...}{a number named indexes}
}
\description{
Helper function to sum over a set of indexes
}
