% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination-status.R
\docType{data}
\name{TERMINATION_STATUS}
\alias{TERMINATION_STATUS}
\title{Solver termination status codes}
\format{An object of class \code{environment} of length 7.}
\usage{
TERMINATION_STATUS
}
\description{
Solver termination status codes
}
\keyword{datasets}
