% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-expression-methods.R
\name{*,MOI_scalar_affine_function,MOI_scalar_affine_function-method}
\alias{*,MOI_scalar_affine_function,MOI_scalar_affine_function-method}
\title{Multiply two scalar affine functions}
\usage{
\S4method{*}{MOI_scalar_affine_function,MOI_scalar_affine_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar affine function}

\item{e2}{A scalar affine function}
}
\description{
Multiply two scalar affine functions
}
