% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{*,MOI_scalar_quadratic_function,numeric-method}
\alias{*,MOI_scalar_quadratic_function,numeric-method}
\title{Multiply a scalar quadratic function and a numeric}
\usage{
\S4method{*}{MOI_scalar_quadratic_function,numeric}(e1, e2)
}
\arguments{
\item{e1}{A scalar quadratic function}

\item{e2}{A scalar numeric}
}
\description{
Multiply a scalar quadratic function and a numeric
}
\keyword{internal}
