% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-expression-methods.R
\name{*,numeric,MOI_scalar_quadratic_function-method}
\alias{*,numeric,MOI_scalar_quadratic_function-method}
\title{Multiply a scalar quadratic function and a numeric}
\usage{
\S4method{*}{numeric,MOI_scalar_quadratic_function}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric}

\item{e2}{A scalar quadratic function}
}
\description{
Multiply a scalar quadratic function and a numeric
}
\keyword{internal}
