% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-class.R
\docType{class}
\name{RMPK_variable-class}
\alias{RMPK_variable-class}
\alias{RMPK_constraint-class}
\alias{RMPK_variable_list-class}
\title{A variable reference}
\description{
A variable reference

A variable reference
}
