% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-interface.R
\docType{data}
\name{RMPSolverInterface}
\alias{RMPSolverInterface}
\title{An interface R6 class for a solver}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RMPSolverInterface
}
\description{
You can also use any other data structure that implements the same methods.
}
\keyword{datasets}
