% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-interface.R
\name{RMPSolverInterface}
\alias{RMPSolverInterface}
\title{An interface R6 class for a solver}
\description{
You can also use any other data structure that implements the same methods.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_variable}{\code{RMPSolverInterface$add_variable()}}
\item \href{#method-add_linear_constraint}{\code{RMPSolverInterface$add_linear_constraint()}}
\item \href{#method-set_linear_objective}{\code{RMPSolverInterface$set_linear_objective()}}
\item \href{#method-add_quadratic_constraint}{\code{RMPSolverInterface$add_quadratic_constraint()}}
\item \href{#method-set_quadratic_objective}{\code{RMPSolverInterface$set_quadratic_objective()}}
\item \href{#method-set_variable_lb}{\code{RMPSolverInterface$set_variable_lb()}}
\item \href{#method-set_variable_ub}{\code{RMPSolverInterface$set_variable_ub()}}
\item \href{#method-nvars}{\code{RMPSolverInterface$nvars()}}
\item \href{#method-nconstraints}{\code{RMPSolverInterface$nconstraints()}}
\item \href{#method-optimize}{\code{RMPSolverInterface$optimize()}}
\item \href{#method-get_variable_value}{\code{RMPSolverInterface$get_variable_value()}}
\item \href{#method-get_variable_dual}{\code{RMPSolverInterface$get_variable_dual()}}
\item \href{#method-get_row_dual}{\code{RMPSolverInterface$get_row_dual()}}
\item \href{#method-set_variable_value}{\code{RMPSolverInterface$set_variable_value()}}
\item \href{#method-get_objective_value}{\code{RMPSolverInterface$get_objective_value()}}
\item \href{#method-get_termination_status}{\code{RMPSolverInterface$get_termination_status()}}
\item \href{#method-get_termination_message}{\code{RMPSolverInterface$get_termination_message()}}
\item \href{#method-format}{\code{RMPSolverInterface$format()}}
\item \href{#method-clone}{\code{RMPSolverInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_variable"></a>}}
\subsection{Method \code{add_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$add_variable(type, lower_bound = -Inf, upper_bound = Inf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_linear_constraint"></a>}}
\subsection{Method \code{add_linear_constraint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$add_linear_constraint(linear_expr, type, rhs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_linear_objective"></a>}}
\subsection{Method \code{set_linear_objective()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$set_linear_objective(linear_expr, sense)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_quadratic_constraint"></a>}}
\subsection{Method \code{add_quadratic_constraint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$add_quadratic_constraint(quadratic_expr, type, rhs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_quadratic_objective"></a>}}
\subsection{Method \code{set_quadratic_objective()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$set_quadratic_objective(quadratic_expr, sense)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_variable_lb"></a>}}
\subsection{Method \code{set_variable_lb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$set_variable_lb(variable_index, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_variable_ub"></a>}}
\subsection{Method \code{set_variable_ub()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$set_variable_ub(variable_index, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nvars"></a>}}
\subsection{Method \code{nvars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$nvars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nconstraints"></a>}}
\subsection{Method \code{nconstraints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$nconstraints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\subsection{Method \code{optimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$optimize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable_value"></a>}}
\subsection{Method \code{get_variable_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_variable_value(var_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable_dual"></a>}}
\subsection{Method \code{get_variable_dual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_variable_dual(var_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_row_dual"></a>}}
\subsection{Method \code{get_row_dual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_row_dual(row_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_variable_value"></a>}}
\subsection{Method \code{set_variable_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$set_variable_value(var_index, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objective_value"></a>}}
\subsection{Method \code{get_objective_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_objective_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_termination_status"></a>}}
\subsection{Method \code{get_termination_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_termination_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_termination_message"></a>}}
\subsection{Method \code{get_termination_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$get_termination_message()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RMPSolverInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
