% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-class.R
\docType{class}
\name{RMPKVariable-class}
\alias{RMPKVariable-class}
\alias{format,RMPKVariable-method}
\alias{print,RMPKVariable-method}
\alias{show,RMPKVariable-method}
\alias{RMPKVariableList-class}
\alias{format,RMPKVariableList-method}
\alias{print,RMPKVariableList-method}
\alias{show,RMPKVariableList-method}
\title{A variable reference}
\usage{
\S4method{format}{RMPKVariable}(x, ...)

\S4method{print}{RMPKVariable}(x, ...)

\S4method{show}{RMPKVariable}(object)

\S4method{format}{RMPKVariableList}(x, ...)

\S4method{print}{RMPKVariableList}(x, ...)

\S4method{show}{RMPKVariableList}(object)
}
\description{
A variable reference
}
