% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-class.R
\docType{class}
\name{RLPVariable-class}
\alias{RLPVariable-class}
\alias{format,RLPVariable-method}
\alias{print,RLPVariable-method}
\alias{show,RLPVariable-method}
\alias{RLPVariableList-class}
\alias{format,RLPVariableList-method}
\alias{print,RLPVariableList-method}
\alias{show,RLPVariableList-method}
\title{A variable reference}
\usage{
\S4method{format}{RLPVariable}(x, ...)

\S4method{print}{RLPVariable}(x, ...)

\S4method{show}{RLPVariable}(object)

\S4method{format}{RLPVariableList}(x, ...)

\S4method{print}{RLPVariableList}(x, ...)

\S4method{show}{RLPVariableList}(object)
}
\description{
A variable reference
}
